requirements.txt
numpy
tensorflow
keras
matplotlib
scikit-learn
opencv-python
Pillow
streamlit